Fix errors resulting from over-long Zaqar queue names
Zaqar queue names can not be longer than 64 characters. To cope
gracefully with this:

- Limit generated queue names to <=64 characters
- Limit user-supplied queue names to 1-64 characters
- Verify the queue name length in the zaqar.queue constraint
- Don't fail to clean up a subscription that failed to create

Closes-Bug: #1655767

+++ heat/engine/resources/openstack/zaqar/subscription.py
@@ -119,12 +119,10 @@ class ZaqarSubscription(resource.Resource):
         subscription.update(self._subscription_data())
 
     def handle_delete(self):
-        try:
+        if self.resource_id is None:
+            return
+        with self.client_plugin().ignore_not_found:
             self._get_subscription().delete()
-        except Exception as ex:
-            self.client_plugin().ignore_not_found(ex)
-        else:
-            return True
 
     def _show_resource(self):
         subscription = self._get_subscription()
